Reflect resource.id in --tag-regexps
we claim to delete based on name/tag, but in reality we only relied on
tags. But some resources are not really tagged by the name, but they
really only use the "id" property so let's include that to our list of
tags to apply the "--tag-regexps" to it as well.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>